Add keyboard support
